updated lists of messages

+++ frontend/src/components/admin/ListTrash.js
@@ -6,8 +6,8 @@ import Loader from '../layout/Loader'
 import Sidebar from './Sidebar'
 import { useAlert } from 'react-alert'
 import { useDispatch, useSelector } from 'react-redux'
-import { deleteInquiry, listInquiry, clearErrors } from '../../actions/inquiryActions'
-import { DELETE_INQUIRY_RESET } from '../../constants/inquiryConstants'
+import { deleteInquiry, updateInquiry, listInquiry, clearErrors } from '../../actions/inquiryActions'
+import { DELETE_INQUIRY_RESET, UPDATE_INQUIRY_RESET } from '../../constants/inquiryConstants'
 
 const ListTrash = ( { history} ) => {
 
@@ -15,7 +15,7 @@ const ListTrash = ( { history} ) => {
     const dispatch = useDispatch();
 
     const { loading, error, inquiries } = useSelector(state => state.listInquiry)
-    const { deleteError, isDeleted } = useSelector(state => state.inquiry)
+    const { deleteError, isUpdated, isDeleted } = useSelector(state => state.inquiry)
 
     useEffect(() => {
         dispatch(listInquiry());
@@ -30,6 +30,15 @@ const ListTrash = ( { history} ) => {
             dispatch(clearErrors())
         }
 
+        if(isUpdated){
+            alert.success('Inquiry has been updated successfully.');
+            history.push('/admin/trash')
+
+            dispatch({
+                type: UPDATE_INQUIRY_RESET
+            })
+        }
+
         if(isDeleted){
             alert.success('Inquiry has been deleted successfully.');
             history.push('/admin/trash')
@@ -38,7 +47,7 @@ const ListTrash = ( { history} ) => {
                 type: DELETE_INQUIRY_RESET
             })
         }
-    }, [dispatch, alert, error, history, isDeleted, deleteError])
+    }, [dispatch, alert, error, history, isDeleted, isUpdated, deleteError])
 
     const setInquiries = () => {
         const data = { 
@@ -86,9 +95,12 @@ const ListTrash = ( { history} ) => {
                     companyName: inquiry.companyName,
                     concernType: String(inquiry.concernType),
                     actions:   <Fragment>
-                                <Link to={`/admin/inquiry/${inquiry._id}`} className='btn btn-primary'>
+                                <Link to={`/admin/inquiry/${inquiry._id}`} className='btn btn-primary py-1 px-2 ml-2'>
                                     <i className='fa fa-eye'></i>
                                 </Link>
+                                <button className="btn btn-secondary py-1 px-2 ml-2" onClick={() => updateInquiryHandler(inquiry._id, "Resolved")}>
+                                    <i className='fa fa-undo'></i>
+                                </button>
                                 <button className="btn btn-danger py-1 px-2 ml-2" onClick={() => deleteInquiryHandler(inquiry._id)}>
                                     <i className='fa fa-trash'></i>
                                 </button>
@@ -100,13 +112,23 @@ const ListTrash = ( { history} ) => {
          return data
     }
 
+    const updateInquiryHandler = (id, inquiryStatus) => { 
+        const formData = new FormData();
+        formData.set('inquiryStatus', inquiryStatus);
+
+        dispatch(updateInquiry(id, formData));
+    }
+
     const deleteInquiryHandler = (id) => {
-        dispatch(deleteInquiry(id))
+
+        if(window.confirm("Are you sure you want to delete this message? This cannot be undone.")){
+            dispatch(deleteInquiry(id))
+        }
     }
 
     return (
         <Fragment>
-            <MetaData title={'Archives'}/>
+            <MetaData title={'Trash'}/>
             <div className="row">
                 <div className="col-4 col-md-2">
                     <Sidebar/>
